Edit Site: Move template creation to the template switcher.
