+++ packages/edit-site/src/components/template-switcher/index.js
@@ -3,11 +3,18 @@
  */
 import { useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
+import { useState, useCallback } from '@wordpress/element';
 import {
 	DropdownMenu,
 	MenuGroup,
 	MenuItemsChoice,
+	MenuItem,
 } from '@wordpress/components';
+
+/**
+ * Internal dependencies
+ */
+import AddTemplate from '../add-template';
 
 export default function TemplateSwitcher( {
 	ids,
@@ -16,6 +23,7 @@ export default function TemplateSwitcher( {
 	isTemplatePart,
 	onActiveIdChange,
 	onActiveTemplatePartIdChange,
+	onAddTemplateId,
 } ) {
 	const { templates, templateParts } = useSelect(
 		( select ) => {
@@ -39,35 +47,53 @@ export default function TemplateSwitcher( {
 		},
 		[ ids, templatePartIds ]
 	);
+	const [ isAddTemplateOpen, setIsAddTemplateOpen ] = useState( false );
 	return (
-		<DropdownMenu
-			icon="layout"
-			label={ __( 'Switch Template' ) }
-			toggleProps={ {
-				children: ( isTemplatePart ? templateParts : templates ).find(
-					( choice ) => choice.value === activeId
-				).label,
-			} }
-			className="edit-site-template-switcher"
-		>
-			{ () => (
-				<>
-					<MenuGroup label={ __( 'Templates' ) }>
-						<MenuItemsChoice
-							choices={ templates }
-							value={ ! isTemplatePart ? activeId : undefined }
-							onSelect={ onActiveIdChange }
-						/>
-					</MenuGroup>
-					<MenuGroup label={ __( 'Template Parts' ) }>
-						<MenuItemsChoice
-							choices={ templateParts }
-							value={ isTemplatePart ? activeId : undefined }
-							onSelect={ onActiveTemplatePartIdChange }
-						/>
-					</MenuGroup>
-				</>
-			) }
-		</DropdownMenu>
+		<>
+			<DropdownMenu
+				icon="layout"
+				label={ __( 'Switch Template' ) }
+				toggleProps={ {
+					children: ( isTemplatePart ? templateParts : templates ).find(
+						( choice ) => choice.value === activeId
+					).label,
+				} }
+				className="edit-site-template-switcher"
+			>
+				{ ( { onClose } ) => (
+					<>
+						<MenuGroup label={ __( 'Templates' ) }>
+							<MenuItemsChoice
+								choices={ templates }
+								value={ ! isTemplatePart ? activeId : undefined }
+								onSelect={ onActiveIdChange }
+							/>
+							<MenuItem
+								icon="plus"
+								onClick={ () => {
+									onClose();
+									setIsAddTemplateOpen( true );
+								} }
+							>
+								{ __( 'New' ) }
+							</MenuItem>
+						</MenuGroup>
+						<MenuGroup label={ __( 'Template Parts' ) }>
+							<MenuItemsChoice
+								choices={ templateParts }
+								value={ isTemplatePart ? activeId : undefined }
+								onSelect={ onActiveTemplatePartIdChange }
+							/>
+						</MenuGroup>
+					</>
+				) }
+			</DropdownMenu>
+			<AddTemplate
+				ids={ ids }
+				onAddTemplateId={ onAddTemplateId }
+				onRequestClose={ useCallback( () => setIsAddTemplateOpen( false ), [] ) }
+				isOpen={ isAddTemplateOpen }
+			/>
+		</>
 	);
 }
